fixing the fix
Setting the bound for the increase in prefered_area as half the initial contractility leads to the increase never happening because contractility is a factor (1+percentage) and the prefered_area values are always much larger (28 and up). The condition makes more sense if it depends on the initial prefered_area.

This still does not exactly replicate the published version! There, the increase in prefered_area is conditional on the contractility not going below its bound, rather than prefered_area itself going above some bound. But with initial*1.8 = 50.4 as an upper bound we get close to the value where the published version cuts off, which is 50.6979